Centralize the interface include/exclude logic
The logic that defined interface aliases for the host name
was separate from that used to define interfaces for the
inter-daemon communication system (oob). Combine the two
and update the interface include/exclude parameters (leaving
deprecated synonyms for the prior forms).
